WebCore: Convert column span from an unsigned short type to an unsigned int type. Fixes a divide-by-zero crash arising from using a zero colspan value coming from a narrow cast of an int to an unsigned short. https://bugs.webkit.org/show_bug.cgi?id=40812
Patch by Abhishek Arya <inferno@chromium.org> on 2010-06-18
Reviewed by Adam Barth.

Test: fast/table/zero-colspan-crash.html

* rendering/RenderTable.h: Change span from unsigned short to unsigned int.
* rendering/RenderTableSection.cpp: Fix a compiler warning with comparing
  unsigned int with signed int. Value of an unsigned int here cannot be
  greater than maximum positive value of a signed int.
(WebCore::RenderTableSection::addCell):

LayoutTests: Tests that we do not crash while rendering a fixed table layout.
https://bugs.webkit.org/show_bug.cgi?id=40812

Patch by Abhishek Arya <inferno@chromium.org> on 2010-06-18
Reviewed by Adam Barth.

* fast/table/zero-colspan-crash-expected.txt: Added.
* fast/table/zero-colspan-crash.html: Added.

Canonical link: https://commits.webkit.org/52461@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@61451 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/rendering/RenderTable.h
@@ -75,7 +75,7 @@ public:
         {
         }
 
-        unsigned short span;
+        unsigned span;
         unsigned width; // the calculated position of the column
     };
